+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## v0.2.0 - 2021-05-21
+
+New:
+- Add `disallowedUrls` option.
+
+Breaking change:
+- `disallowedBotNames` option becomes `disallowedCrawlers`.
+
+
 ## v0.1.0 - 2020-09-02
 
 - Initial verion